+++ VoidEngine/VoidEngine/ThirdPersonCamera.hpp
@@ -5,7 +5,7 @@ class ThirdPersonCamera : public Camera
 {
 public:
 	ThirdPersonCamera();
-	virtual void	handleMouseMovement(float x, float y);
+	virtual void	handleMouseMovement(int x, int y);
 	virtual void	handleKeyboard(char input);
 	void			setSpeed(float speed);
