+++ VoidEngine/VoidEngine/TopDownCamera.hpp
@@ -5,7 +5,7 @@ class TopDownCamera : public Camera
 {
 public:
 	TopDownCamera();
-	virtual void	handleMouseMovement(float x, float y);
+	virtual void	handleMouseMovement(int x, int y);
 	virtual void	handleKeyboard(char input);
 	void			setSpeed(float speed);
